+++ starsky/starskycore/Interfaces/ISearchSuggest.cs
@@ -6,7 +6,7 @@ namespace starskycore.Interfaces
 	public interface ISearchSuggest
 	{
 		IEnumerable<string> SearchSuggest(string query);
-
+		IEnumerable<KeyValuePair<string, int>> GetAllSuggestions();
 		List<KeyValuePair<string,int>> Inflate();
 	}
 }